requirement.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Deprecated==1.2.6
diff-match-patch==20181111
get==2019.4.13
gitdb2==2.0.5
GitPython==3.0.2
idna==2.8
importlib-metadata==0.19
more-itertools==7.2.0
-e git+git@github.com:OpenPecha/openpecha.git@9ffecfb1bd59b2be7ba8245db00ebba2624b148d#egg=openpoti
packaging==19.1
pluggy==0.12.0
post==2019.4.13
public==2019.4.13
py==1.8.0
PyGithub==1.43.8
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.1
PyYAML==5.1.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
smmap2==2.0.5
tqdm==4.35.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
atomicwrites==1.3.0
attrs==19.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Deprecated==1.2.6
diff-match-patch==20181111
docutils==0.15.2
get==2019.4.13
gitdb2==2.0.5
GitPython==3.0.2
idna==2.8
importlib-metadata==0.19
more-itertools==7.2.0
-e git+git@github.com:OpenPecha/openpecha.git@5f2cb480a8e41971dd04bd2269a412f1cfa12229#egg=openpoti
packaging==19.1
pkginfo==1.5.0.1
pluggy==0.12.0
post==2019.4.13
public==2019.4.13
py==1.8.0
PyGithub==1.43.8
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.1
PyYAML==5.1.2
query-string==2019.4.13
readme-renderer==24.0
request==2019.4.13
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
smmap2==2.0.5
tqdm==4.35.0
twine==2.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
zipp==0.6.0
